perf: defer optional editor, draft sync, analytics, and library menu code
