+++ src/components/Navbar.tsx
@@ -1,5 +1,10 @@
 import * as React from 'react'
 import { twMerge } from 'tailwind-merge'
+const LazyLibrariesMenuContent = React.lazy(() =>
+  import('./LibrariesMenuContent').then((module) => ({
+    default: module.LibrariesMenuContent,
+  })),
+)
 const LazyAiDock = React.lazy(() =>
   import('./AiDock').then((m) => ({ default: m.AiDock })),
 )
@@ -42,15 +47,6 @@ import { AiDockButton, SearchButton } from './SearchButton'
 import { BrandContextMenu } from './BrandContextMenu'
 import { useSearchContext } from '~/contexts/SearchContext'
 import { useLibrariesOverlay } from '~/contexts/LibrariesOverlayContext'
-import { publicLibraries, type LibrarySlim } from '~/libraries'
-import {
-  categoryLabels,
-  categoryOrder,
-  categoryTextColor,
-  libraryCategories,
-  type LibraryCategory,
-} from '~/libraries/categories'
-import { fallbackLibraryIcon, libraryIcons } from '~/libraries/icons'
 import { GithubIcon } from '~/components/icons/GithubIcon'
 import {
   Dropdown,
@@ -336,69 +332,6 @@ const NAV_GROUPS = [
     },
   },
 ] as const satisfies readonly NavMenuGroup[]
-
-function getLibraryDisplayName(library: LibrarySlim) {
-  return library.name.replace(/^TanStack\s+/, '')
-}
-
-type LibraryMenuEntry = {
-  id: string
-  name: string
-  to: string
-  icon: IconComponent
-  /** `group-hover/lib:text-category-*` — recolors the icon to its category. */
-  iconHoverColor: string
-}
-
-// Full static class strings (Tailwind can't see composed names) mapping each
-// category to the hover color applied to a library's icon in the mega-menu.
-const categoryIconHoverColor: Record<LibraryCategory, string> = {
-  framework: 'group-hover/lib:text-category-framework',
-  data: 'group-hover/lib:text-category-data',
-  ui: 'group-hover/lib:text-category-ui',
-  performance: 'group-hover/lib:text-category-performance',
-  tooling: 'group-hover/lib:text-category-tooling',
-}
-
-type LibraryMenuColumn = {
-  category: LibraryCategory
-  label: string
-  colorClass: string
-  libraries: LibraryMenuEntry[]
-}
-
-/**
- * The Libraries mega-menu as five category columns (Framework, Data & State,
- * UI & UX, Performance, Tooling), built from the canonical `libraryCategories`
- * taxonomy. Iterating `libraryCategories` preserves the intended per-category
- * order; only public, navigable libraries are shown.
- */
-function getLibraryCategoryColumns(): LibraryMenuColumn[] {
-  const byCategory = new Map<LibraryCategory, LibraryMenuEntry[]>(
-    categoryOrder.map((category) => [category, []]),
-  )
-
-  for (const [id, category] of Object.entries(libraryCategories)) {
-    const library = publicLibraries.find((lib) => lib.id === id)
-    if (!library || !library.to) continue
-    byCategory.get(category)?.push({
-      id: library.id,
-      name: getLibraryDisplayName(library),
-      to: library.to,
-      icon: libraryIcons[library.id] ?? fallbackLibraryIcon,
-      iconHoverColor: categoryIconHoverColor[category],
-    })
-  }
-
-  return categoryOrder
-    .map((category) => ({
-      category,
-      label: categoryLabels[category],
-      colorClass: categoryTextColor[category],
-      libraries: byCategory.get(category) ?? [],
-    }))
-    .filter((column) => column.libraries.length > 0)
-}
 
 function AiDockMount() {
   const { isAiDockOpen } = useSearchContext()
@@ -760,6 +693,7 @@ function DesktopNavTrigger({
   onDismiss: () => void
   onResetDismissed: () => void
 }) {
+  const [hasLoadedLibraries, setHasLoadedLibraries] = React.useState(false)
   const { openLibraries } = useLibrariesOverlay()
   const triggerClassName = twMerge(
     'ts-mega-trigger inline-flex items-center gap-1 rounded-md px-2 py-2 text-xs font-medium min-[1120px]:gap-1.5 min-[1120px]:px-3 min-[1120px]:text-[13px]',
@@ -775,8 +709,14 @@ function DesktopNavTrigger({
       onAuxClick={(event) => {
         if (event.button === 1) onDismiss()
       }}
+      onPointerEnter={() => {
+        if (group.key === 'libraries') setHasLoadedLibraries(true)
+      }}
       onPointerLeave={onResetDismissed}
-      onFocusCapture={onResetDismissed}
+      onFocusCapture={() => {
+        onResetDismissed()
+        if (group.key === 'libraries') setHasLoadedLibraries(true)
+      }}
     >
       {group.key === 'libraries' ? (
         <button
@@ -812,7 +752,11 @@ function DesktopNavTrigger({
           <span>{group.label}</span>
         </button>
       )}
-      <DesktopNavDropdown group={group} onNavigate={onDismiss} />
+      <DesktopNavDropdown
+        group={group}
+        onNavigate={onDismiss}
+        loadLibraries={hasLoadedLibraries}
+      />
     </div>
   )
 }
@@ -820,9 +764,11 @@ function DesktopNavTrigger({
 function DesktopNavDropdown({
   group,
   onNavigate,
+  loadLibraries,
 }: {
   group: NavMenuGroup
   onNavigate: () => void
+  loadLibraries: boolean
 }) {
   return (
     <div className="ts-mega-dropdown">
@@ -836,11 +782,15 @@ function DesktopNavDropdown({
             'min-[1120px]:px-[43px] min-[1120px]:pt-12 min-[1120px]:pb-[38px]',
         )}
       >
-        <MegaMenuContent
-          group={group}
-          onNavigate={onNavigate}
-          variant="desktop"
-        />
+        {group.key !== 'libraries' || loadLibraries ? (
+          <MegaMenuContent
+            group={group}
+            onNavigate={onNavigate}
+            variant="desktop"
+          />
+        ) : (
+          <LibrariesMenuFallback />
+        )}
       </div>
     </div>
   )
@@ -988,6 +938,17 @@ function MobileNavigation({
   )
 }
 
+function LibrariesMenuFallback() {
+  return (
+    <div
+      role="status"
+      className="flex min-h-40 w-[min(880px,calc(100vw-3rem))] items-center justify-center text-text-secondary"
+    >
+      Loading libraries…
+    </div>
+  )
+}
+
 function MegaMenuContent({
   group,
   onNavigate,
@@ -998,7 +959,11 @@ function MegaMenuContent({
   variant: 'desktop' | 'mobile'
 }) {
   if (group.key === 'libraries') {
-    return <LibrariesMenuContent onNavigate={onNavigate} variant={variant} />
+    return (
+      <React.Suspense fallback={<LibrariesMenuFallback />}>
+        <LazyLibrariesMenuContent onNavigate={onNavigate} variant={variant} />
+      </React.Suspense>
+    )
   }
 
   if (group.key === 'learn') {
@@ -1077,188 +1042,6 @@ function MegaMenuContent({
         <MenuRail rail={group.rail} onNavigate={onNavigate} variant={variant} />
       ) : null}
     </div>
-  )
-}
-
-function LibrariesMenuContent({
-  onNavigate,
-  variant,
-}: {
-  onNavigate: () => void
-  variant: 'desktop' | 'mobile'
-}) {
-  const { openLibraries } = useLibrariesOverlay()
-  const columns = getLibraryCategoryColumns()
-
-  const allLibraries = (
-    <Button
-      type="button"
-      onClick={() => {
-        openLibraries()
-        onNavigate()
-      }}
-      variant="subtle-link"
-      color="gray"
-      className={twMerge(
-        'group/all gap-1.5 rounded-lg px-[9px] py-2 text-ds-mono-xs focus:text-text-primary focus:outline-none',
-        variant === 'desktop' &&
-          'min-[1120px]:gap-[7px] min-[1120px]:rounded-[10px] min-[1120px]:px-[11px] min-[1120px]:py-2.5 min-[1120px]:text-[14px]',
-      )}
-    >
-      <GridFourIcon
-        className={twMerge(
-          'size-4',
-          variant === 'desktop' && 'min-[1120px]:size-[19px]',
-        )}
-      />
-      Browse all libraries
-      <ArrowRightIcon
-        className={twMerge(
-          'size-3.5 transition-transform group-hover/all:translate-x-0.5',
-          variant === 'desktop' && 'min-[1120px]:size-[17px]',
-        )}
-      />
-    </Button>
-  )
-
-  if (variant === 'mobile') {
-    return (
-      <div className="grid gap-4">
-        {columns.map((column) => (
-          <LibraryCategoryColumn
-            key={column.category}
-            column={column}
-            onNavigate={onNavigate}
-            variant="mobile"
-          />
-        ))}
-        {allLibraries}
-      </div>
-    )
-  }
-
-  return (
-    <div className="flex flex-col gap-4 min-[1120px]:gap-5">
-      <div className="flex items-start gap-9 min-[1120px]:gap-12">
-        {columns.map((column) => (
-          <LibraryCategoryColumn
-            key={column.category}
-            column={column}
-            onNavigate={onNavigate}
-            variant="desktop"
-          />
-        ))}
-      </div>
-      <div className="flex justify-center border-t border-border-subtle pt-1.5 min-[1120px]:pt-2">
-        {allLibraries}
-      </div>
-    </div>
-  )
-}
-
-function LibraryCategoryColumn({
-  column,
-  onNavigate,
-  variant,
-}: {
-  column: LibraryMenuColumn
-  onNavigate: () => void
-  variant: 'desktop' | 'mobile'
-}) {
-  return (
-    <div
-      className={twMerge(
-        'flex flex-col',
-        variant === 'desktop'
-          ? 'w-[120px] gap-4 min-[1120px]:w-36 min-[1120px]:gap-5'
-          : 'gap-1',
-      )}
-    >
-      <div
-        className={`pl-[9px] font-ds-mono uppercase ${
-          variant === 'desktop' ? 'text-ds-mono-sm' : 'text-ds-mono-xs'
-        } ${variant === 'desktop' ? 'min-[1120px]:pl-[11px]' : ''} ${column.colorClass}`}
-      >
-        {column.label}
-      </div>
-      <div className="flex flex-col items-start gap-1">
-        {column.libraries.map((library) => (
-          <LibraryMenuRow
-            key={library.id}
-            library={library}
-            onNavigate={onNavigate}
-            variant={variant}
-          />
-        ))}
-      </div>
-    </div>
-  )
-}
-
-function LibraryMenuRow({
-  library,
-  onNavigate,
-  variant,
-}: {
-  library: LibraryMenuEntry
-  onNavigate: () => void
-  variant: 'desktop' | 'mobile'
-}) {
-  const Icon = library.icon
-  const external = library.to.startsWith('http')
-  const className = twMerge(
-    // Light mode: an "elevated white" hover — a bright-white pill lifted off the
-    // glass with a soft shadow + hairline ring (contrast via depth, not value).
-    // Dark mode keeps the subtle white/4% (pressed 12%) overlay, no shadow/ring.
-    'group/lib flex items-center gap-2 rounded-[14px] py-2 pl-[9px] pr-4 text-text-secondary transition-[color,background-color,box-shadow] hover:bg-white hover:text-text-primary hover:shadow-sm hover:ring-1 hover:ring-black/5 focus:bg-white focus:text-text-primary focus:shadow-sm focus:ring-1 focus:ring-black/5 focus:outline-none active:bg-white dark:hover:bg-text-primary/[0.04] dark:hover:shadow-none dark:hover:ring-0 dark:focus:bg-text-primary/[0.04] dark:focus:shadow-none dark:focus:ring-0 dark:active:bg-text-primary/[0.12]',
-    variant === 'desktop'
-      ? 'h-[38px] min-[1120px]:h-[46px] min-[1120px]:gap-2.5 min-[1120px]:rounded-[17px] min-[1120px]:pl-[11px] min-[1120px]:pr-[18px]'
-      : 'py-2.5',
-  )
-  const content = (
-    <>
-      {/* Plain template string: the category hover color is a `text-*` utility
-          and twMerge would drop it against a base color. */}
-      <Icon
-        className={`size-5 shrink-0 transition-colors ${
-          variant === 'desktop' ? 'min-[1120px]:size-6' : ''
-        } ${library.iconHoverColor}`}
-      />
-      <span
-        className={twMerge(
-          'whitespace-nowrap font-ds-display text-[16px] tracking-[0.32px]',
-          variant === 'desktop' &&
-            'min-[1120px]:text-[19px] min-[1120px]:tracking-[0.38px]',
-        )}
-      >
-        {library.name}
-      </span>
-    </>
-  )
-
-  if (external) {
-    return (
-      <a
-        href={library.to}
-        target="_blank"
-        rel="noopener noreferrer"
-        className={className}
-        onClick={onNavigate}
-      >
-        {content}
-      </a>
-    )
-  }
-
-  return (
-    <Link
-      to={library.to}
-      onClick={onNavigate}
-      preload="intent"
-      className={className}
-    >
-      {content}
-    </Link>
   )
 }
